+++ modesolver/__init__.py
@@ -11,13 +11,13 @@ This package currently provides:
     https://doi.org/10.1109/JLT.2008.923643
 """
 
-__version__ = "2.0.0"
+__version__ = "3.0.0"
 __author__ = "Thomas E. Murphy"
 
 from .wgmodes import wgmodes
 from .wgmodes_yee import wgmodes_yee
 from .svmodes import svmodes
 from .geometry import waveguidemesh, waveguidemeshfull, fibermesh, stretchmesh, padmesh, trimmesh
-from .postprocess import efields, poynting, unfold
+from .postprocess import collocate, poynting, unfold
 
-__all__ = ["wgmodes", "wgmodes_yee", "svmodes", "waveguidemesh", "waveguidemeshfull", "fibermesh", "stretchmesh", "padmesh", "trimmesh", "efields", "poynting", "unfold"]
+__all__ = ["wgmodes", "wgmodes_yee", "svmodes", "waveguidemesh", "waveguidemeshfull", "fibermesh", "stretchmesh", "padmesh", "trimmesh", "collocate", "poynting", "unfold"]