+++ modesolver/postprocess/__init__.py
@@ -1,3 +1,3 @@
-from .efields import efields
+from .collocate import collocate
 from .poynting import poynting
 from .unfold import unfold